+++ src/solver/assemble_initial_condition.rs
@@ -3,7 +3,7 @@ use rand::seq::SliceRandom;
 
 /// Make an initial condition of the appropriate size `grid_size` from prescribed data.
 /// Fill everything with the state `fill`, except for the indices in the hashmap.
-pub fn assemble_initial_condition(fill: usize, different: HashMap<u64, usize>, graph_size: u64) -> Vec<usize> {
+pub fn assemble_initial_condition(fill: usize, different: HashMap<usize, usize>, graph_size: usize) -> Vec<usize> {
     let mut initial_condition: Vec<usize> = Vec::new();
 
     for i in 0..graph_size {
@@ -18,7 +18,7 @@ pub fn assemble_initial_condition(fill: usize, different: HashMap<u64, usize>, g
 /// Make an initial condition of the appropriate size `grid_size` by sampling from a distribution.
 /// A random entry from the hash set `states` will be chosen. Weights can be assigned by repeating a
 /// particular state in the `states` vector.
-pub fn assemble_random_initial_condition(states: Vec<usize>, grid_size: u64) -> Vec<usize> {
+pub fn assemble_random_initial_condition(states: Vec<usize>, grid_size: usize) -> Vec<usize> {
     let mut rng = rand::thread_rng();
 
     let mut initial_condition: Vec<usize> = Vec::new();
